Fixing some data types

+++ ports/riot/mpconfigport.h
@@ -24,6 +24,7 @@
  * THE SOFTWARE.
  */
 #include <stdlib.h>
+#include <stdint.h>
 
 #include "irq.h"
 
@@ -96,8 +97,8 @@
 
 #define MICROPY_MODULE_FROZEN_STR       (1)
 
-typedef int mp_int_t;       // must be pointer size
-typedef unsigned mp_uint_t; // must be pointer size
+typedef intptr_t mp_int_t;       // must be pointer size
+typedef uintptr_t mp_uint_t; // must be pointer size
 typedef long mp_off_t;
 
 #define MICROPY_BEGIN_ATOMIC_SECTION()     irq_disable()
